+++ app/routes/admin/messages.js
@@ -1,13 +1,17 @@
-import classic from 'ember-classic-decorator';
 import Route from '@ember/routing/route';
+import EmberTableRouteMixin from 'open-event-frontend/mixins/ember-table-route';
 
-@classic
-export default class MessagesRoute extends Route {
+export default class extends Route.extend(EmberTableRouteMixin) {
   titleToken() {
     return this.l10n.t('Messages');
   }
 
-  model() {
-    return this.store.query('message-setting', {});
+  async model(params) {
+    let queryString = {
+      'page[size]'  : params.per_page || 100,
+      'page[number]' : params.page || 1
+    };
+    queryString = this.applySortFilters(queryString, params);
+    return this.asArray(this.store.query('message-setting', queryString));
   }
 }
